feat: no edition of quantities in basket
I chose to keep quantities in basket items, but make them nullable even for saved and confirmed baskets. That way, external apps can still specify quantities if they need to using the HTTP API.
The quantities are displayed (but not editable) in the basket if at least one of them is present. If not, they are hidden.