chore: remove moment with a custom webpack config
Isntead of manually deleting moment, this uses an extra webpack config to do the job.
Even if this is not an official Angular package, it is the recommended solution by the CLI team.

It gets rid of the warning we currently have in the build `Module not found: Error: Can't resolve './locale' in '/Users/ced-pro/Code/ninjasquad/inra/rare-basket/frontend/node_modules/moment/src/lib/locale'`.
And it works on any platform, whereas the current postinstall script works only for developers in Unix env.

It also works better in fact, as this is the build size before this commit:
```
284K 29 mai 12:40 order-order-module-es2015.d9b36ce871ea583b2031.js
289K 29 mai 12:40 order-order-module-es5.d9b36ce871ea583b2031.js
```
vs with this solution:
```
212K 29 mai 12:43 order-order-module-es2015.4db8030ba4385170b34a.js
217K 29 mai 12:43 order-order-module-es5.4db8030ba4385170b34a.js
```