feat: use accessionHolder field in basket
The `contactEmail` field has been changed to an `accessionHolder` field, directly referencing the accession holder name.
To reflect that the unicity constraint on the name is now among all the accession holders, and not only inside a GRC.

The testing data have also been updated to reflect the testing data of `data-discovery`. The accesion holders now have real names.
A new accession holder has been added with `contect22/password`, as we have 4 accession holders in the testing data.